refactor: extract event_types.rs from daemon lib.rs (Sprint 8)
Extract 210+ lines of type definitions and channel helpers from the
daemon monolith (lib.rs) into a dedicated event_types.rs module:

- OutputId struct
- Event enum (17 variants)
- DaemonNodeEvent enum (7 variants)
- AdoraEvent enum (4 variants)
- RunStatus enum
- ZenohOutbound struct
- send_with_timestamp / send_drop_with_timestamp functions
- NODE_EVENT_CHANNEL_CAPACITY / CONTROL_EVENT_HEADROOM constants
- InterDaemonEvent re-export

lib.rs: 4251 -> 4039 lines (-212)
event_types.rs: 231 lines (new)

Pure mechanical extraction — zero behavior change. All 18 daemon
tests pass. This is the first step of the daemon module split.
